Add {% include code solution_name='my_solution.cc' start=5 end=12 %}

+++ _includes/source_code/code/25-PDP/triangle/triangle_efficient.cc
@@ -5,7 +5,7 @@
 const size_t MAXN = 1000;
 
 int T[MAXN][MAXN];
-int max_path[MAXN][MAXN];
+long max_path[MAXN][MAXN];
 int N;
 
 int main() { 
@@ -18,22 +18,22 @@ int main() {
    }
    fclose(fi);
    
-   max_path[0][0] = T[0][0];
+   max_path[0][0] = (long)T[0][0];
    for (int i = 1; i < N; ++i) {
       for (int j = 0; j <= i; ++j) {
-         max_path[i][j] = T[i][j];
+	 max_path[i][j] = (long)T[i][j];
          if (j == 0) max_path[i][j] += max_path[i - 1][j];
          else max_path[i][j] += std::max(max_path[i - 1][j - 1], max_path[i - 1][j]);
       }
    }
    
-   int max_path_to_bottom = max_path[N-1][0];
+   long max_path_to_bottom = max_path[N-1][0];
    for (int i = 0; i < N; ++i) {
       max_path_to_bottom = std::max(max_path[N-1][i], max_path_to_bottom);
    }
 
    FILE *fo = fopen("triangle.out", "w");
-   fprintf(fo, "%d\n", max_path_to_bottom);
+   fprintf(fo, "%ld\n", max_path_to_bottom);
    fclose(fo);
    return 0;
 }
